+++ pages/contact-us.tsx
@@ -1,0 +1,18 @@
+// components
+import ContactUsForm from "src/containers/ContactUs";
+
+// layout
+import AuthLayout from "src/layouts/AuthLayout";
+
+
+function ContactUs() {
+
+
+  return (
+    <AuthLayout>
+      <ContactUsForm/>
+    </AuthLayout>
+  );
+}
+
+export default ContactUs;