+++ pages/register.tsx
@@ -1,4 +1,6 @@
-import { useState } from "react";
+import { useRouter } from "next/router";
+import { useFormik } from "formik";
+import { toast } from "react-toastify";
 
 //layout
 import AuthLayout from "src/layouts/AuthLayout";
@@ -8,18 +10,40 @@ import AuthFormLayout from "src/layouts/AuthFormLayout";
 import Button from "src/elements/Button";
 import Input from "src/elements/Input";
 
+// helpers
+import auth from "src/helpers/api/auth";
+
+// types
+import { InterfaceRegisterData } from "src/types/api/auth";
+
 function Register() {
-  const [userName, setUserName] = useState<string>("");
+  const router = useRouter();
+
+  const { handleChange, values, handleSubmit } = useFormik({
+    initialValues: {
+      email: "",
+      password: "",
+      name: "",
+    },
+    onSubmit: (data: InterfaceRegisterData) => {
+      auth.register(data)
+        .then(() => {
+          router.push("/login");
+          toast.success("New User has been created");
+        })
+        .catch(e => toast.error(e.details))
+    },
+  });
+
+  const { email, password, name } = values;
 
 
   return (
     <AuthLayout>
-      <AuthFormLayout>
-        <Input placeholder="email" value={userName} onChange={e => setUserName(e.target.value)}/>
-        <Input placeholder="first name" value={userName} onChange={e => setUserName(e.target.value)}/>
-        <Input placeholder="last name" value={userName} onChange={e => setUserName(e.target.value)}/>
-        <Input placeholder="user name" value={userName} onChange={e => setUserName(e.target.value)}/>
-        <Input placeholder="password" value={userName} onChange={e => setUserName(e.target.value)}/>
+      <AuthFormLayout onSubmit={handleSubmit}>
+        <Input placeholder="Email" value={email} onChange={handleChange} name="email"/>
+        <Input placeholder="Name" value={name} onChange={handleChange} name="name"/>
+        <Input placeholder="password" value={password} onChange={handleChange} name="password"/>
         <Button>
           Login
         </Button>