+++ src/helpers/graphcms/home.ts
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+
+const home = {
+  getHomePageData: gql `
+    query HomePageData {
+      howItWorksPlural {
+        title
+        icon {
+          url
+        }
+        description
+      }
+      homePageTops {
+        title
+        description
+        image {
+          url
+        }
+      }
+    }
+  `,
+};
+
+export default home;